image-generator/requirements.txt
diffusers==0.27.2
transformers==4.40.1
accelerate==0.29.2
safetensors==0.4.2
huggingface-hub==0.23.1
scipy==1.13.0
Pillow==10.3.0
fastapi==0.110.0
uvicorn[standard]==0.29.0
numpy==1.26.4
python-multipart==0.0.9
aiofiles==23.2.0

